Refactor import paths for NamedArray and relocate array.py
Adjusted import paths for NamedArray to simplify references in multiple files (__init__.py, v1/variable.py, v0/variable.py). Moved array.py from base directory to schemas directory to further improve the organization of the code. The Compressors import in the relocated array.py file was realigned to reflect this change.
